[Marcus.min] 방탈출 인증 step1 & step2 제출합니다 (#26)
* feat: step1 & step2

* test: mocking 테스트 추가 (practice)

* refactor: AuthService에서 MemberDao가 아닌 MemberService를 활용하기

* feat: AuthInterceptor 추가

* refactor: AuthenticatedId -> AuthenticatedUsername 변경

* refactor: Exception 추가

* refactor: Lombok Annotation 추가

* refactor: Exception 추가

* refactor: Scheme 재정의

* refactor: DTO & Entity Lombok 적용 & 스키마 변경

* refactor: 관계 재정의에 의한 Dao 재정의

* test: Dao 테스트코드 추가

* refactor: 관계 재정의에 의한 Service 재정의

* test: Service mocking test 추가

* refactor: 관계 재정의에 의한 Controller 재정의

* refactor: token principle 변경 (username -> id)

* test: Test code 추가 및 정리

* refactor: use checkWrongPassword method

+++ src/main/java/nextstep/reservation/ReservationController.java
@@ -1,5 +1,10 @@
 package nextstep.reservation;
 
+import nextstep.auth.AuthenticatedMember;
+import nextstep.exceptions.exception.notFound.ReservationForbiddenException;
+import nextstep.member.Member;
+import nextstep.schedule.Schedule;
+import nextstep.schedule.ScheduleService;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
@@ -10,33 +15,49 @@ import java.util.List;
 @RequestMapping("/reservations")
 public class ReservationController {
 
-    public final ReservationService reservationService;
+    private final ReservationService reservationService;
+    private final ScheduleService scheduleService;
 
-    public ReservationController(ReservationService reservationService) {
+    public ReservationController(ReservationService reservationService,
+            ScheduleService scheduleService) {
         this.reservationService = reservationService;
+        this.scheduleService = scheduleService;
     }
 
     @PostMapping
-    public ResponseEntity createReservation(@RequestBody ReservationRequest reservationRequest) {
-        Long id = reservationService.create(reservationRequest);
+    public ResponseEntity<Object> createReservation(
+            @RequestBody ReservationRequest reservationRequest,
+            @AuthenticatedMember Member member
+    ) {
+        Schedule schedule = scheduleService.findById(reservationRequest.getScheduleId());
+        Long id = reservationService.create(member, schedule);
+        if (!reservationRequest.getName().equals(member.getName())) {
+            throw new ReservationForbiddenException("예약자가 일치해야만 예약을 생성할 수 있습니다.");
+        }
         return ResponseEntity.created(URI.create("/reservations/" + id)).build();
     }
 
     @GetMapping
-    public ResponseEntity readReservations(@RequestParam Long themeId, @RequestParam String date) {
-        List<Reservation> results = reservationService.findAllByThemeIdAndDate(themeId, date);
-        return ResponseEntity.ok().body(results);
+    public ResponseEntity<List<Reservation>> readReservations(
+            @RequestParam Long themeId,
+            @RequestParam String date,
+            @AuthenticatedMember Member member
+    ) {
+        List<Reservation> reservations = reservationService.findAllByThemeIdAndDate(themeId, date);
+        return ResponseEntity.ok().body(reservations);
     }
 
     @DeleteMapping("/{id}")
-    public ResponseEntity deleteReservation(@PathVariable Long id) {
+    public ResponseEntity<Object> deleteReservation(
+            @PathVariable Long id,
+            @AuthenticatedMember Member member
+    ) {
+        Reservation reservation = reservationService.findById(id);
+        if (!reservation.getMember().getId().equals(member.getId())) {
+            throw new ReservationForbiddenException("예약당사자만 예약을 삭제할 수 있습니다.");
+        }
         reservationService.deleteById(id);
-
         return ResponseEntity.noContent().build();
     }
 
-    @ExceptionHandler(Exception.class)
-    public ResponseEntity onException(Exception e) {
-        return ResponseEntity.badRequest().build();
-    }
 }
